+++ x/authz/client/cli/tx.go
@@ -67,7 +67,10 @@ func NewCmdExecAuthorization() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			grantee := clientCtx.GetFromAddress()
+			grantee, err := clientCtx.AddressCodec.BytesToString(clientCtx.GetFromAddress())
+			if err != nil {
+				return err
+			}
 
 			if offline, _ := cmd.Flags().GetBool(flags.FlagOffline); offline {
 				return errors.New("cannot broadcast tx during offline mode")
@@ -106,13 +109,13 @@ Examples:
 				return err
 			}
 
-			if strings.EqualFold(args[0], clientCtx.GetFromAddress().String()) {
-				return errors.New("grantee and granter should be different")
-			}
-
-			grantee, err := clientCtx.AddressCodec.StringToBytes(args[0])
+			grantee := args[0]
+			granter, err := clientCtx.AddressCodec.BytesToString(clientCtx.GetFromAddress())
 			if err != nil {
 				return err
+			}
+			if strings.EqualFold(args[0], granter) {
+				return errors.New("grantee and granter should be different")
 			}
 
 			var authorization authz.Authorization
@@ -230,7 +233,7 @@ Examples:
 				return err
 			}
 
-			msg, err := authz.NewMsgGrant(clientCtx.GetFromAddress(), grantee, authorization, expire)
+			msg, err := authz.NewMsgGrant(granter, grantee, authorization, expire)
 			if err != nil {
 				return err
 			}